Added PeakDriver.PcanBasicNet; Changed a bit PeakDriver.Core;

+++ PeakDriver.Core/Components/PeakMessage.cs
@@ -3,7 +3,7 @@
     public struct PeakMessage
     {
         #region Constructors
-        public PeakMessage(uint timestampt, byte[] data)
+        public PeakMessage(ulong timestampt, byte[] data)
         {
             Timestampt = timestampt;
             Data = data;
@@ -11,7 +11,7 @@
         #endregion
 
         #region Properties
-        public uint Timestampt { get; }
+        public ulong Timestampt { get; }
         public byte[] Data { get; }
         #endregion
     }
